requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2
psycopg2-binary
pytz
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0